+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+
+- Dropped the manual page for `rsync-ssl`. It documents a helper script that
+  is not part of this binary, so `unpin man rsync-ssl` was answering for a
+  command there is no way to run.